refactor: 「ユーザの指定を持って、インポート情報を変更する」というメソッドを、ImportDiagramCandidateへ持っていく。
