+++ src/application/service/diagram/import/DiagramImportService.ts
@@ -1,5 +1,5 @@
 import { DiagramImportProgressStep } from "@/domain/diagram/import/progress/DiagramImportProgressStep";
-import { DiagramImportError } from "~/domain/diagram/import/progress/DiagramImportError";
+import { DiagramImportError } from "@/domain/diagram/import/progress/DiagramImportError";
 import DiagramImportProgressEvent from "@/domain/diagram/import/progress/DiagramImportProgressEvent";
 import StorageRepository from "@/domain/storage/StorageRepository";
 import FileSystemRepository from "@/domain/filesystem/FileSystemRepository";
@@ -8,10 +8,8 @@ import Diagram from "@/domain/diagram/Diagram";
 import Product from "@/domain/product/Product";
 import UserArrangeOfImportDiagramSetting from "@/domain/diagram/import/userarrange/UserArrangeOfImportDiagramSetting";
 import NameOfColided from "@/domain/diagram/import/userarrange/NameOfColided";
-import { BehaviorWhenNameColide } from "@/domain/diagram/import/userarrange/BehavioWhenNameColide";
-import Resource from "~/domain/resource/Resource";
-import MaybeImportDiagram from "~/domain/diagram/import/MaybeImportDiagram";
-import ImportDiagramCandidate from "~/domain/diagram/import/ImportDiagramCandidate";
+import MaybeImportDiagram from "@/domain/diagram/import/MaybeImportDiagram";
+import ImportDiagramCandidate from "@/domain/diagram/import/ImportDiagramCandidate";
 
 export default class DiagramImportService {
     constructor(
@@ -93,14 +91,12 @@ export default class DiagramImportService {
         product: Product
     ): ImportDiagramCandidate | null {
         const colidedNames = candidate.analyzeColideNameOf(product);
+        if (colidedNames.isEmpty()) return candidate;
 
-        let userArrange = colidedNames;
-        if (!colidedNames.isEmpty()) {
-            userArrange = confirmeUserArrange(colidedNames);
-            if (userArrange.isEmpty()) return null;
-        }
+        const userArrange = confirmeUserArrange(colidedNames);
+        if (userArrange.isEmpty()) return null;
 
-        return this.arrangeImportDiagram(userArrange, candidate, product);
+        return candidate.arrangeImportDiagram(userArrange, product);
     }
 
     private analyzeColideNameOf(candidate: ImportDiagramCandidate, product: Product): UserArrangeOfImportDiagramSetting {
@@ -118,57 +114,6 @@ export default class DiagramImportService {
             .map(r => NameOfColided.prototypeResourceOf(r));
 
         return new UserArrangeOfImportDiagramSetting(diagram.name, colidedName, sameResources);
-    }
-
-    private arrangeImportDiagram(
-        userArrange: UserArrangeOfImportDiagramSetting,
-        candidate: ImportDiagramCandidate,
-        product: Product
-    ): ImportDiagramCandidate | null {
-        let modified = candidate;
-
-        for (const colidedResourceName of userArrange.resourceNamesOfColided) {
-            const targetResouce = modified.useResources
-                .find(r => r.resourceId === colidedResourceName.sourceId) as Resource;
-            const sameResource = product.resources.getSameOf(targetResouce) as Resource;
-
-            const behavior = colidedResourceName.behavior;
-            if (behavior === BehaviorWhenNameColide.既存) {
-                const replacedDiagram = modified.diagram.replaceOf(targetResouce, sameResource);
-                const removedResources = modified.useResources
-                    .filter(r => r.resourceId !== targetResouce.resourceId);
-                modified = new ImportDiagramCandidate(replacedDiagram, removedResources);
-            }
-            if (behavior === BehaviorWhenNameColide.置換) {
-                const replacedIdResource = targetResouce.renewId(sameResource.resourceId);
-                const replacedExportedResources = modified.useResources
-                    .filter(r => r.resourceId !== targetResouce.resourceId);
-                replacedExportedResources.push(replacedIdResource);
-                const replacedDiagram = modified.diagram
-                    .replaceOf(targetResouce, replacedIdResource);
-                modified = new ImportDiagramCandidate(replacedDiagram, replacedExportedResources);
-            }
-            if (behavior === BehaviorWhenNameColide.別名) {
-                const renamedResource = targetResouce
-                    .withName(colidedResourceName.destinationName);
-                const replacedExportedResources = modified.useResources
-                    .filter(r => r.resourceId !== targetResouce.resourceId);
-                replacedExportedResources.push(renamedResource);
-                modified = new ImportDiagramCandidate(modified.diagram, replacedExportedResources);
-            }
-        }
-
-        if (userArrange.isColidedDiagramName()) {
-            const colidedDiagramName = userArrange.diagramNamesOfColided as NameOfColided;
-            if (colidedDiagramName.behavior === BehaviorWhenNameColide.既存) return null; // 入力からは入ってこない前提。「既存」というなら「Importしない」と同義。
-            if (colidedDiagramName.behavior === BehaviorWhenNameColide.別名) {
-                const renamedDiagram = modified.diagram
-                    .renameOf(colidedDiagramName.destinationName)
-                modified = new ImportDiagramCandidate(renamedDiagram, modified.useResources);
-            }
-        }
-
-        return modified;
     }
 
     private mergeOf(candidate: ImportDiagramCandidate, product: Product): Product {
