+++ src/domain/diagram/import/ImportDiagramCandidate.ts
@@ -3,6 +3,7 @@ import NameOfColided from "@/domain/diagram/import/userarrange/NameOfColided";
 import Resource from "@/domain/resource/Resource";
 import Diagram from "@/domain/diagram/Diagram";
 import Product from "@/domain/product/Product";
+import { BehaviorWhenNameColide } from "@/domain/diagram/import/userarrange/BehavioWhenNameColide";
 
 export default class ImportDiagramCandidate {
     constructor(
@@ -43,4 +44,52 @@ export default class ImportDiagramCandidate {
 
         return new UserArrangeOfImportDiagramSetting(diagram.name, colidedName, sameResources);
     }
+
+    public arrangeImportDiagram(
+        userArrange: UserArrangeOfImportDiagramSetting,
+        product: Product
+    ): ImportDiagramCandidate | null {
+        let modifiedDiagram = this.diagram;
+        let modifiedResources = this.useResources;
+
+        for (const colidedResourceName of userArrange.resourceNamesOfColided) {
+            const targetResouce = modifiedResources
+                .find(r => r.resourceId === colidedResourceName.sourceId) as Resource;
+            const sameResource = product.resources.getSameOf(targetResouce) as Resource;
+
+            const behavior = colidedResourceName.behavior;
+            if (behavior === BehaviorWhenNameColide.既存) {
+                modifiedDiagram = modifiedDiagram.replaceOf(targetResouce, sameResource);
+                modifiedResources = modifiedResources
+                    .filter(r => r.resourceId !== targetResouce.resourceId);
+            }
+            if (behavior === BehaviorWhenNameColide.置換) {
+                const replacedIdResource = targetResouce.renewId(sameResource.resourceId);
+                modifiedResources = modifiedResources
+                    .filter(r => r.resourceId !== targetResouce.resourceId);
+                modifiedResources.push(replacedIdResource);
+                modifiedDiagram = modifiedDiagram
+                    .replaceOf(targetResouce, replacedIdResource);
+            }
+            if (behavior === BehaviorWhenNameColide.別名) {
+                const renamedResource = targetResouce
+                    .withName(colidedResourceName.destinationName);
+                modifiedResources = modifiedResources
+                    .filter(r => r.resourceId !== targetResouce.resourceId);
+                modifiedResources.push(renamedResource);
+            }
+        }
+
+        if (userArrange.isColidedDiagramName()) {
+            const colidedDiagramName = userArrange.diagramNamesOfColided as NameOfColided;
+            if (colidedDiagramName.behavior === BehaviorWhenNameColide.既存) return null; // 入力からは入ってこない前提。「既存」というなら「Importしない」と同義。
+            if (colidedDiagramName.behavior === BehaviorWhenNameColide.別名) {
+                modifiedDiagram = modifiedDiagram
+                    .renameOf(colidedDiagramName.destinationName);
+            }
+        }
+
+        return new ImportDiagramCandidate(modifiedDiagram, modifiedResources);
+    }
+
 }
